+++ src/modules/network/index.ts
@@ -1,0 +1,10 @@
+/*!
+ * Copyright (c) Grab Taxi Holdings PTE LTD (GRAB)
+ *
+ * This source code is licensed under the MIT license found in the LICENSE file in the root
+ * directory of this source tree.
+ */
+
+export { NetworkModule } from './NetworkModule';
+export { SendRequestSchema, SendResponseSchema, SendResultSchema } from './schemas';
+export type { SendRequest, SendResponse, SendResult } from './types';